Rename plexlint to jotspot
I was approached by a Plex employee and urged to rename, because the "plex" in the name "plexlint" implies that it is an official Plex tool, which it is not.

For those wondering, a jot is literally "the smallest bit." So, to spot a jot means to find the smallest bit, which is what we're trying to do: find the smallest, most obscure problems in a large collection of files.